Remove irdi_parser dependency
Previously, we used the [irdi_parser] project as
a dependency to parse IRDI identifiers.
However, we had problems getting a fully automatic
building of a Docker image to run, as it failed
at this dependency for some reason.

We therefore decide to remove the dependency, but
still refer to the source of the regular
expression used to match the IRDI.

[irdi_parser](https://github.com/moritzsommer/irdi-parser)